fix: migrate to dom observer
